requirements.txt
dagster
dagster-webserver
pandas
discord.py
fastf1
pytz
scikit-learn
pyodbc
SQLAlchemy
seaborn
dagster-mysql
tensorflow
mysql-connector-python
matplotlib
numpy